update gps code

+++ Project1/3년차/gps/gps_with_gyro2.cpp
@@ -358,11 +358,11 @@ double get_gpsData(char *chrBuf, int num)
 double get_groundSpeed(char *chrBuf, int num)
 {
     double gh; double gy; double gv;
-    signed int tmp[8];
+    signed short tmp[8];
     unsigned char i;
 
     for(i=0;i<8;i++){
-        tmp[i] = (signed int)chrBuf[i+2];
+        tmp[i] = (signed short)chrBuf[i+2];
     }
 
     if (num==1){
